Moves all shape code into 1 file called shapes.py, updates 2d box.
